+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Master
 
 - Add target and slot support
+- Add ref
+- Move transform system to C
 
 ## v0.0.4.302